Wrote the framework code for getting organelle efficiency updates to
the GUI, also did a lot of script bindings refactoring
