+++ scripts/microbe_stage/microbe_editor/energy_balance.as
@@ -1,0 +1,9 @@
+//! \brief Calculates the energy balance for a cell with the given organelles
+//!
+//! Used by the editor to collect info to send to the GUI
+#include "calculate_effectiveness.as"
+
+void calculateEnergyBalanceWithOrganelles()
+{
+
+}
