+++ scripts/microbe_stage/microbe_stage.levgm
@@ -41,6 +41,8 @@ o GameModule "microbe_stage"{
         microbe_editor/microbe_editor.as
         microbe_editor/microbe_editor_hud.as
         microbe_editor/setup.as
+        microbe_editor/calculate_effectiveness.as
+        microbe_editor/energy_balance.as
         
         // TODO: This should be easy to move to another module
         utility.as
