+++ src/scripting/script_initializer.cpp
@@ -7,202 +7,14 @@
 #include "general/hex.h"
 #include "general/locked_map.h"
 #include "general/properties_component.h"
-#include "general/timed_life_system.h"
-#include "generated/cell_stage_world.h"
-#include "generated/microbe_editor_world.h"
-#include "microbe_stage/organelle_types.h"
-#include "microbe_stage/patch.h"
-#include "microbe_stage/player_microbe_control.h"
-#include "microbe_stage/simulation_parameters.h"
 #include "microbe_stage/species.h"
-#include "microbe_stage/species_name_controller.h"
 
 #include "ThriveGame.h"
 
-
 #include <Script/Bindings/BindHelpers.h>
-#include <Script/Bindings/StandardWorldBindHelper.h>
 #include <Script/ScriptExecutor.h>
 
-
 using namespace thrive;
-// ------------------------------------ //
-// Proxies and helpers
-
-//! This is safe to use as long as scripting module is valid
-static asITypeInfo* mapWrapperTypeInfo = nullptr;
-
-CScriptArray*
-    patchMapGetPatchesWrapper(const PatchMap& self)
-{
-    const auto& patches = self.getPatches();
-
-    LEVIATHAN_ASSERT(
-        mapWrapperTypeInfo, "map wrapper type info is not retrieved");
-
-    CScriptArray* array = CScriptArray::Create(mapWrapperTypeInfo);
-
-    if(!array)
-        return nullptr;
-
-    array->Reserve(static_cast<asUINT>(patches.size()));
-
-    for(auto iter = patches.begin(); iter != patches.end(); ++iter) {
-        Patch* tmp = iter->second.get();
-        array->InsertLast(&tmp);
-    }
-
-    return array;
-}
-
-CScriptArray*
-    patchGetNeighboursWrapper(const Patch& self)
-{
-    const auto& patches = self.getNeighbours();
-    return Leviathan::ConvertIteratorToASArray(patches.begin(), patches.end(),
-        Leviathan::ScriptExecutor::Get()->GetASEngine());
-}
-
-
-//! Wrapper for TJsonRegistry::getSize
-template<class RegistryT>
-uint64_t
-    getSizeWrapper(RegistryT* self)
-{
-    return static_cast<uint64_t>(self->getSize());
-}
-
-//! Wrapper for TJsonRegistry::getTypeData
-template<class RegistryT, class ReturnedT>
-const ReturnedT*
-    getTypeDataWrapper(RegistryT* self, uint64_t id)
-{
-    return &self->getTypeData(id);
-}
-
-// Wrappers for registerSimulationDataAndJsons
-
-SpeciesNameController*
-    getNameWrapper()
-{
-    return &SimulationParameters::speciesNameController;
-}
-
-TJsonRegistry<Compound>*
-    getCompoundRegistryWrapper()
-{
-    return &SimulationParameters::compoundRegistry;
-}
-
-TJsonRegistry<BioProcess>*
-    getBioProcessRegistryWrapper()
-{
-    return &SimulationParameters::bioProcessRegistry;
-}
-
-TJsonRegistry<Biome>*
-    getBiomeRegistryWrapper()
-{
-    return &SimulationParameters::biomeRegistry;
-}
-
-TJsonRegistry<OrganelleType>*
-    getOrganelleRegistryWrapper()
-{
-    return &SimulationParameters::organelleRegistry;
-}
-
-class ScriptSpawnerWrapper {
-public:
-    //! \note Caller must have incremented ref count already on func
-    ScriptSpawnerWrapper(asIScriptFunction* func) : m_func(func)
-    {
-        if(!m_func)
-            throw std::runtime_error("no func given for ScriptSpawnerWrapper");
-    }
-
-    ~ScriptSpawnerWrapper()
-    {
-        m_func->Release();
-    }
-
-    ObjectID
-        run(CellStageWorld& world, Float3 pos)
-    {
-        ScriptRunningSetup setup;
-        auto result = Leviathan::ScriptExecutor::Get()->RunScript<ObjectID>(
-            m_func, nullptr, setup, &world, pos);
-
-        if(result.Result != SCRIPT_RUN_RESULT::Success) {
-
-            LOG_ERROR("Failed to run Wrapped SpawnSystem function");
-            // This makes the spawn system just ignore the return value
-            return NULL_OBJECT;
-        }
-
-        return result.Value;
-    }
-
-    asIScriptFunction* m_func;
-};
-
-SpawnerTypeId
-    addSpawnTypeProxy(SpawnSystem* self,
-        asIScriptFunction* func,
-        double spawnDensity,
-        double spawnRadius)
-{
-    auto wrapper = std::make_shared<ScriptSpawnerWrapper>(func);
-
-    return self->addSpawnType(
-        [=](CellStageWorld& world, Float3 pos) -> ObjectID {
-            return wrapper->run(world, pos);
-        },
-        spawnDensity, spawnRadius);
-}
-
-
-class WorldEffectScript : public WorldEffect {
-public:
-    //! \note Caller must have incremented ref count already on func
-    WorldEffectScript(asIScriptFunction* func) : m_func(func)
-    {
-        if(!m_func)
-            throw std::runtime_error("no func given to WorldEffectScript");
-    }
-
-    ~WorldEffectScript()
-    {
-        m_func->Release();
-    }
-
-    void
-        onTimePassed(double elapsed, long double totalTimePassed) override
-    {
-        double totalConverted = static_cast<double>(totalTimePassed);
-
-        ScriptRunningSetup setup;
-        auto result = Leviathan::ScriptExecutor::Get()->RunScript<void>(
-            m_func, nullptr, setup, m_world, elapsed, totalConverted);
-
-        if(result.Result != SCRIPT_RUN_RESULT::Success) {
-
-            LOG_ERROR("Failed to run WorldEffectScript function");
-        }
-    }
-
-    asIScriptFunction* m_func;
-};
-
-
-void
-    registerEffectProxy(TimedWorldOperations& self,
-        const std::string& name,
-        asIScriptFunction* func)
-{
-    self.registerEffect(name, std::make_unique<WorldEffectScript>(func));
-}
-
 // ------------------------------------ //
 bool
     registerLockedMap(asIScriptEngine* engine)
@@ -278,539 +90,6 @@ bool
     return true;
 }
 
-//! Helper for registerSimulationDataAndJsons
-template<class RegistryT, class ReturnedT>
-bool
-    registerJsonRegistry(asIScriptEngine* engine,
-        const char* classname,
-        const std::string& returnedTypeName)
-{
-    if(engine->RegisterObjectType(classname, 0, asOBJ_REF | asOBJ_NOCOUNT) <
-        0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod(classname, "uint64 getSize()",
-           asFUNCTION(getSizeWrapper<RegistryT>), asCALL_CDECL_OBJFIRST) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod(classname,
-           ("const " + returnedTypeName + "@ getTypeData(uint64 id)").c_str(),
-           asFUNCTION((getTypeDataWrapper<RegistryT, ReturnedT>)),
-           asCALL_CDECL_OBJFIRST) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod(classname,
-           "uint64 getTypeId(const string &in internalName)",
-           asMETHOD(RegistryT, getTypeId), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod(classname,
-           "const string& getInternalName(uint64 id)",
-           asMETHOD(RegistryT, getInternalName), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    return true;
-}
-
-//! Helper for registerJsonregistryHeldTypes
-template<class RegistryT>
-bool
-    registerRegistryHeldHelperBases(asIScriptEngine* engine,
-        const char* classname)
-{
-    if(engine->RegisterObjectType(classname, 0, asOBJ_REF | asOBJ_NOCOUNT) <
-        0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectProperty(
-           classname, "uint64 id", asOFFSET(RegistryT, id)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty(classname, "string displayName",
-           asOFFSET(RegistryT, displayName)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty(classname, "const string internalName",
-           asOFFSET(RegistryT, internalName)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    return true;
-}
-
-bool
-    registerJsonRegistryHeldTypes(asIScriptEngine* engine)
-{
-    if(!registerRegistryHeldHelperBases<Compound>(engine, "Compound"))
-        return false;
-
-    if(!registerRegistryHeldHelperBases<BioProcess>(engine, "BioProcess"))
-        return false;
-
-    if(!registerRegistryHeldHelperBases<Biome>(engine, "Biome"))
-        return false;
-
-    if(!registerRegistryHeldHelperBases<OrganelleType>(engine, "OrganelleType"))
-        return false;
-
-    // Compound specific properties //
-    // ------------------------------------ //
-    // Compound
-    if(engine->RegisterObjectProperty(
-           "Compound", "double volume", asOFFSET(Compound, volume)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty(
-           "Compound", "bool isCloud", asOFFSET(Compound, isCloud)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty(
-           "Compound", "bool isUseful", asOFFSET(Compound, isUseful)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("Compound", "bool isEnvironmental",
-           asOFFSET(Compound, isEnvironmental)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty(
-           "Compound", "Float4 colour", asOFFSET(Compound, colour)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    // ------------------------------------ //
-    // Biome
-    // TODO: bind new light properties
-    if(engine->RegisterObjectProperty("Biome", "const string background",
-           asOFFSET(Biome, background)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    // TODO: reference counting for these
-    if(engine->RegisterObjectType(
-           "BiomeCompoundData", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("Biome",
-           "const BiomeCompoundData@ getCompound(uint64 type) const",
-           asMETHOD(Biome, getCompound), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("Biome",
-           "BiomeCompoundData@ getCompound(uint64 type)",
-           asMETHOD(Biome, getCompound), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("Biome",
-           "array<uint64>@ getCompoundKeys() const",
-           asMETHOD(Biome, getCompoundKeys), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("BiomeCompoundData", "uint amount",
-           asOFFSET(BiomeCompoundData, amount)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("BiomeCompoundData", "double density",
-           asOFFSET(BiomeCompoundData, density)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("BiomeCompoundData", "double dissolved",
-           asOFFSET(BiomeCompoundData, dissolved)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectType("ChunkData", 0, asOBJ_REF | asOBJ_NOCOUNT) <
-        0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("Biome",
-           "const ChunkData& getChunk(uint64 type) const",
-           asMETHOD(Biome, getChunk), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("Biome",
-           "array<uint64>@ getChunkKeys() const", asMETHOD(Biome, getChunkKeys),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty(
-           "ChunkData", "string name", asOFFSET(ChunkData, name)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty(
-           "ChunkData", "double density", asOFFSET(ChunkData, density)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty(
-           "ChunkData", "bool dissolves", asOFFSET(ChunkData, dissolves)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty(
-           "ChunkData", "uint radius", asOFFSET(ChunkData, radius)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty(
-           "ChunkData", "uint mass", asOFFSET(ChunkData, mass)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty(
-           "ChunkData", "uint size", asOFFSET(ChunkData, size)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("ChunkData", "double ventAmount",
-           asOFFSET(ChunkData, ventAmount)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty(
-           "ChunkData", "double damages", asOFFSET(ChunkData, damages)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("ChunkData", "bool deleteOnTouch",
-           asOFFSET(ChunkData, deleteOnTouch)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("ChunkData", "double chunkScale",
-           asOFFSET(ChunkData, chunkScale)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectType(
-           "ChunkCompoundData", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("ChunkData",
-           "const ChunkCompoundData& getCompound(uint64 type) const",
-           asMETHOD(ChunkData, getCompound), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("ChunkData",
-           "array<uint64>@ getCompoundKeys() const",
-           asMETHOD(ChunkData, getCompoundKeys), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("ChunkData",
-           "const uint64 getMeshListSize() const",
-           asMETHOD(ChunkData, getMeshListSize), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("ChunkData",
-           "const string getMesh(uint64 index) const",
-           asMETHOD(ChunkData, getMesh), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("ChunkData",
-           "const string getTexture(uint64 index) const",
-           asMETHOD(ChunkData, getTexture), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("ChunkCompoundData", "double amount",
-           asOFFSET(ChunkCompoundData, amount)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("ChunkCompoundData", "string name",
-           asOFFSET(ChunkCompoundData, name)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    // ------------------------------------ //
-    // OrganelleType
-    if(engine->RegisterObjectProperty("OrganelleType", "const string name",
-           asOFFSET(OrganelleType, name)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("OrganelleType", "const string gene",
-           asOFFSET(OrganelleType, gene)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("OrganelleType", "const string mesh",
-           asOFFSET(OrganelleType, mesh)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("OrganelleType", "const string texture",
-           asOFFSET(OrganelleType, texture)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("OrganelleType", "const float mass",
-           asOFFSET(OrganelleType, mass)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("OrganelleType",
-           "const float chanceToCreate",
-           asOFFSET(OrganelleType, chanceToCreate)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("OrganelleType",
-           "const float prokaryoteChance",
-           asOFFSET(OrganelleType, prokaryoteChance)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("OrganelleType",
-           "const array<string>@ getComponentKeys() const",
-           asMETHOD(OrganelleType, getComponentKeys), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("OrganelleType",
-           "const array<string>@ getComponentParameterKeys(string component) "
-           "const",
-           asMETHOD(OrganelleType, getComponentParameterKeys),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("OrganelleType",
-           "const double getComponentParameterAsDouble(string component, "
-           "string parameter) const",
-           asMETHOD(OrganelleType, getComponentParameterAsDouble),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("OrganelleType",
-           "const string getComponentParameterAsString(string component, "
-           "string parameter) const",
-           asMETHOD(OrganelleType, getComponentParameterAsString),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("OrganelleType",
-           "const array<string>@ getProcessKeys() const",
-           asMETHOD(OrganelleType, getProcessKeys), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("OrganelleType",
-           "const float getProcessTweakRate(string process) const",
-           asMETHOD(OrganelleType, getProcessTweakRate), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("OrganelleType",
-           "const array<Int2>@ getHexes() const",
-           asMETHOD(OrganelleType, getHexes), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("OrganelleType",
-           "const array<string>@ getInitialCompositionKeys() const",
-           asMETHOD(OrganelleType, getInitialCompositionKeys),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_ASSUMED_SIZE_T;
-    if(engine->RegisterObjectMethod("OrganelleType",
-           "const double getInitialComposition(string compound) const",
-           asMETHOD(OrganelleType, getInitialComposition),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("OrganelleType", "const int mpCost",
-           asOFFSET(OrganelleType, mpCost)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    return true;
-}
-
-bool
-    registerSimulationDataAndJsons(asIScriptEngine* engine)
-{
-
-    if(engine->RegisterObjectType(
-           "SimulationParameters", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!registerJsonRegistryHeldTypes(engine))
-        return false;
-
-    if(engine->RegisterObjectType(
-           "SpeciesNameController", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("SpeciesNameController",
-           "array<string>@ getVowelPrefixes()",
-           asMETHOD(SpeciesNameController, getVowelPrefixes),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("SpeciesNameController",
-           "array<string>@ getConsonantPrefixes()",
-           asMETHOD(SpeciesNameController, getConsonantPrefixes),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("SpeciesNameController",
-           "array<string>@ getVowelCofixes()",
-           asMETHOD(SpeciesNameController, getVowelCofixes),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("SpeciesNameController",
-           "array<string>@ getConsonantCofixes()",
-           asMETHOD(SpeciesNameController, getConsonantCofixes),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("SpeciesNameController",
-           "array<string>@ getSuffixes()",
-           asMETHOD(SpeciesNameController, getSuffixes), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("SpeciesNameController",
-           "array<string>@ getConsonantSuffixes()",
-           asMETHOD(SpeciesNameController, getConsonantSuffixes),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("SpeciesNameController",
-           "array<string>@ getVowelSuffixes()",
-           asMETHOD(SpeciesNameController, getVowelSuffixes),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("SpeciesNameController",
-           "array<string>@ getPrefixCofix()",
-           asMETHOD(SpeciesNameController, getPrefixCofix),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!registerJsonRegistry<TJsonRegistry<Compound>, Compound>(
-           engine, "TJsonRegistryCompound", "Compound")) {
-        return false;
-    }
-
-    if(!registerJsonRegistry<TJsonRegistry<BioProcess>, BioProcess>(
-           engine, "TJsonRegistryBioProcess", "BioProcess")) {
-        return false;
-    }
-
-    if(!registerJsonRegistry<TJsonRegistry<Biome>, Biome>(
-           engine, "TJsonRegistryBiome", "Biome")) {
-        return false;
-    }
-
-    if(!registerJsonRegistry<TJsonRegistry<OrganelleType>, OrganelleType>(
-           engine, "TJsonRegistryOrganelleType", "OrganelleType")) {
-        return false;
-    }
-
-
-    if(engine->SetDefaultNamespace("SimulationParameters") < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterGlobalFunction(
-           "SpeciesNameController@ speciesNameController()",
-           asFUNCTION(getNameWrapper), asCALL_CDECL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterGlobalFunction(
-           "TJsonRegistryCompound@ compoundRegistry()",
-           asFUNCTION(getCompoundRegistryWrapper), asCALL_CDECL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterGlobalFunction(
-           "TJsonRegistryBioProcess@ bioProcessRegistry()",
-           asFUNCTION(getBioProcessRegistryWrapper), asCALL_CDECL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterGlobalFunction("TJsonRegistryBiome@ biomeRegistry()",
-           asFUNCTION(getBiomeRegistryWrapper), asCALL_CDECL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterGlobalFunction(
-           "TJsonRegistryOrganelleType@ organelleRegistry()",
-           asFUNCTION(getOrganelleRegistryWrapper), asCALL_CDECL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->SetDefaultNamespace("") < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    return true;
-}
 
 bool
     registerSpecies(asIScriptEngine* engine)
@@ -917,514 +196,6 @@ bool
            "Species", "int32 generation", asOFFSET(Species, generation)) < 0) {
         ANGELSCRIPT_REGISTERFAIL;
     }
-
-    return true;
-}
-
-
-static uint16_t ProcessorComponentTYPEProxy =
-    static_cast<uint16_t>(ProcessorComponent::TYPE);
-static uint16_t CompoundVenterTYPEProxy =
-    static_cast<uint16_t>(CompoundVenterComponent::TYPE);
-static uint16_t EngulfableComponentTYPEProxy =
-    static_cast<uint16_t>(EngulfableComponent::TYPE);
-static uint16_t DamageOnTouchComponentTYPEProxy =
-    static_cast<uint16_t>(DamageOnTouchComponent::TYPE);
-static uint16_t SpawnedComponentTYPEProxy =
-    static_cast<uint16_t>(SpawnedComponent::TYPE);
-static uint16_t AgentCloudComponentTYPEProxy =
-    static_cast<uint16_t>(AgentCloudComponent::TYPE);
-static uint16_t CompoundCloudComponentTYPEProxy =
-    static_cast<uint16_t>(CompoundCloudComponent::TYPE);
-static uint16_t MembraneComponentTYPEProxy =
-    static_cast<uint16_t>(MembraneComponent::TYPE);
-static uint16_t FluidEffectComponentTYPEProxy =
-    static_cast<uint16_t>(FluidEffectComponent::TYPE);
-static uint16_t CompoundBagComponentTYPEProxy =
-    static_cast<uint16_t>(CompoundBagComponent::TYPE);
-static uint16_t CompoundAbsorberComponentTYPEProxy =
-    static_cast<uint16_t>(CompoundAbsorberComponent::TYPE);
-static uint16_t TimedLifeComponentTYPEProxy =
-    static_cast<uint16_t>(TimedLifeComponent::TYPE);
-static uint16_t AgentPropertiesTYPEProxy =
-    static_cast<uint16_t>(AgentProperties::TYPE);
-
-//! Helper for bindThriveComponentTypes
-bool
-    bindComponentTypeId(asIScriptEngine* engine,
-        const char* name,
-        uint16_t* value)
-{
-    if(engine->SetDefaultNamespace(name) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterGlobalProperty("const uint16 TYPE", value) < 0) {
-
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->SetDefaultNamespace("") < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    return true;
-}
-
-bool
-    bindThriveComponentTypes(asIScriptEngine* engine)
-{
-
-    if(engine->RegisterObjectType(
-           "ProcessorComponent", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!bindComponentTypeId(
-           engine, "ProcessorComponent", &ProcessorComponentTYPEProxy))
-        return false;
-
-    if(engine->RegisterObjectMethod("ProcessorComponent",
-           "ProcessorComponent& opAssign(const ProcessorComponent &in other)",
-           asMETHODPR(ProcessorComponent, operator=,(const ProcessorComponent&),
-               ProcessorComponent&),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("ProcessorComponent",
-           "void setCapacity(BioProcessId id, double capacity)",
-           asMETHOD(ProcessorComponent, setCapacity), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("ProcessorComponent",
-           "double getCapacity(BioProcessId id)",
-           asMETHOD(ProcessorComponent, getCapacity), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-    // ------------------------------------ //
-    if(engine->RegisterObjectType(
-           "CompoundVenterComponent", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!bindComponentTypeId(
-           engine, "CompoundVenterComponent", &CompoundVenterTYPEProxy))
-        return false;
-
-    if(engine->RegisterObjectMethod("CompoundVenterComponent",
-           "float getVentAmount()",
-           asMETHOD(CompoundVenterComponent, getVentAmount),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("CompoundVenterComponent",
-           "void setVentAmount(float amount)",
-           asMETHOD(CompoundVenterComponent, setVentAmount),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("CompoundVenterComponent",
-           "bool getDoDissolve()",
-           asMETHOD(CompoundVenterComponent, getDoDissolve),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("CompoundVenterComponent",
-           "void setDoDissolve(bool dissolve)",
-           asMETHOD(CompoundVenterComponent, setDoDissolve),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    // ------------------------------------ //
-    if(engine->RegisterObjectType(
-           "EngulfableComponent", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!bindComponentTypeId(
-           engine, "EngulfableComponent", &EngulfableComponentTYPEProxy))
-        return false;
-
-    if(engine->RegisterObjectMethod("EngulfableComponent", "float getSize()",
-           asMETHOD(EngulfableComponent, getSize), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("EngulfableComponent",
-           "void setSize(float size)", asMETHOD(EngulfableComponent, setSize),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    // ------------------------------------ //
-    if(engine->RegisterObjectType(
-           "DamageOnTouchComponent", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!bindComponentTypeId(
-           engine, "DamageOnTouchComponent", &DamageOnTouchComponentTYPEProxy))
-        return false;
-
-    if(engine->RegisterObjectMethod("DamageOnTouchComponent",
-           "double getDamage()", asMETHOD(DamageOnTouchComponent, getDamage),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("DamageOnTouchComponent",
-           "void setDamage(double damage)",
-           asMETHOD(DamageOnTouchComponent, setDamage), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-    if(engine->RegisterObjectMethod("DamageOnTouchComponent",
-           "bool getDeletes()", asMETHOD(DamageOnTouchComponent, getDeletes),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("DamageOnTouchComponent",
-           "void setDeletes(bool deletes)",
-           asMETHOD(DamageOnTouchComponent, setDeletes), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    // ------------------------------------ //
-    if(engine->RegisterObjectType(
-           "SpawnedComponent", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!bindComponentTypeId(
-           engine, "SpawnedComponent", &SpawnedComponentTYPEProxy))
-        return false;
-
-    // ------------------------------------ //
-    if(engine->RegisterObjectType(
-           "AgentCloudComponent", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!bindComponentTypeId(
-           engine, "AgentCloudComponent", &AgentCloudComponentTYPEProxy))
-        return false;
-
-    // ------------------------------------ //
-    if(engine->RegisterObjectType(
-           "CompoundCloudComponent", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!bindComponentTypeId(
-           engine, "CompoundCloudComponent", &CompoundCloudComponentTYPEProxy))
-        return false;
-
-    // ------------------------------------ //
-    if(engine->RegisterObjectType(
-           "FluidEffectComponent", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!bindComponentTypeId(
-           engine, "FluidEffectComponent", &FluidEffectComponentTYPEProxy))
-        return false;
-
-    // ------------------------------------ //
-    if(engine->RegisterObjectType(
-           "MembraneComponent", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!bindComponentTypeId(
-           engine, "MembraneComponent", &MembraneComponentTYPEProxy))
-        return false;
-
-    if(engine->RegisterObjectMethod("MembraneComponent",
-           "void setColour(const Float4 &in colour)",
-           asMETHOD(MembraneComponent, setColour), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("MembraneComponent",
-           "void setHealthFraction(float value)",
-           asMETHOD(MembraneComponent, setHealthFraction),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-
-    if(engine->RegisterEnum("MEMBRANE_TYPE") < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    ANGELSCRIPT_REGISTER_ENUM_VALUE(MEMBRANE_TYPE, MEMBRANE);
-    ANGELSCRIPT_REGISTER_ENUM_VALUE(MEMBRANE_TYPE, DOUBLEMEMBRANE);
-    ANGELSCRIPT_REGISTER_ENUM_VALUE(MEMBRANE_TYPE, WALL);
-    ANGELSCRIPT_REGISTER_ENUM_VALUE(MEMBRANE_TYPE, CHITIN);
-
-    if(engine->RegisterObjectMethod("MembraneComponent",
-           "MEMBRANE_TYPE getMembraneType() const",
-           asMETHOD(MembraneComponent, getMembraneType), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("MembraneComponent",
-           "void setMembraneType(MEMBRANE_TYPE type)",
-           asMETHOD(MembraneComponent, setMembraneType), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("MembraneComponent",
-           "Float4 getColour() const", asMETHOD(MembraneComponent, getColour),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("MembraneComponent", "void clear()",
-           asMETHOD(MembraneComponent, clear), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("MembraneComponent",
-           "float calculateEncompassingCircleRadius() const",
-           asMETHOD(MembraneComponent, calculateEncompassingCircleRadius),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("MembraneComponent",
-           "Float3 GetExternalOrganelle(double x, double y)",
-           asMETHOD(MembraneComponent, GetExternalOrganelle),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("MembraneComponent",
-           "void sendOrganelles(double x, double y)",
-           asMETHOD(MembraneComponent, sendOrganelles), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("MembraneComponent",
-           "bool removeSentOrganelle(double x, double y)",
-           asMETHOD(MembraneComponent, removeSentOrganelle),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    // ------------------------------------ //
-
-    if(engine->RegisterObjectType(
-           "CompoundBagComponent", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!bindComponentTypeId(
-           engine, "CompoundBagComponent", &CompoundBagComponentTYPEProxy))
-        return false;
-
-    if(engine->RegisterObjectMethod("CompoundBagComponent",
-           "double getCompoundAmount(CompoundId compound)",
-           asMETHOD(CompoundBagComponent, getCompoundAmount),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("CompoundBagComponent",
-           "double takeCompound(CompoundId compound, double to_take)",
-           asMETHOD(CompoundBagComponent, takeCompound), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("CompoundBagComponent",
-           "void giveCompound(CompoundId compound, double amount)",
-           asMETHOD(CompoundBagComponent, giveCompound), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("CompoundBagComponent",
-           "void setCompound(CompoundId compound, double amount)",
-           asMETHOD(CompoundBagComponent, setCompound), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("CompoundBagComponent",
-           "double getPrice(CompoundId compound)",
-           asMETHOD(CompoundBagComponent, getPrice), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("CompoundBagComponent",
-           "double getUsedLastTime(CompoundId compound)",
-           asMETHOD(CompoundBagComponent, getUsedLastTime),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("CompoundBagComponent",
-           "double storageSpace",
-           asOFFSET(CompoundBagComponent, storageSpace)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("CompoundBagComponent",
-           "double storageSpaceOccupied",
-           asOFFSET(CompoundBagComponent, storageSpaceOccupied)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    // ------------------------------------ //
-    // CompoundAbsorberComponent
-    if(engine->RegisterObjectType(
-           "CompoundAbsorberComponent", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!bindComponentTypeId(engine, "CompoundAbsorberComponent",
-           &CompoundAbsorberComponentTYPEProxy))
-        return false;
-
-    if(engine->RegisterObjectMethod("CompoundAbsorberComponent",
-           "void enable()", asMETHOD(CompoundAbsorberComponent, enable),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("CompoundAbsorberComponent",
-           "void disable()", asMETHOD(CompoundAbsorberComponent, disable),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-
-    if(engine->RegisterObjectMethod("CompoundAbsorberComponent",
-           "void setGrabScale(float scale)",
-           asMETHOD(CompoundAbsorberComponent, setGrabScale),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-
-    if(engine->RegisterObjectMethod("CompoundAbsorberComponent",
-           "array<CompoundId>@ getAbsorbedCompounds()",
-           asMETHOD(CompoundAbsorberComponent, getAbsorbedCompounds),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-
-    if(engine->RegisterObjectMethod("CompoundAbsorberComponent",
-           "float absorbedCompoundAmount(CompoundId compound)",
-           asMETHOD(CompoundAbsorberComponent, absorbedCompoundAmount),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("CompoundAbsorberComponent",
-           "void setAbsorbtionCapacity(double capacity)",
-           asMETHOD(CompoundAbsorberComponent, setAbsorbtionCapacity),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("CompoundAbsorberComponent",
-           "void setCanAbsorbCompound(CompoundId id, bool canAbsorb)",
-           asMETHOD(CompoundAbsorberComponent, setCanAbsorbCompound),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    // ------------------------------------ //
-    if(engine->RegisterObjectType(
-           "TimedLifeComponent", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!bindComponentTypeId(
-           engine, "TimedLifeComponent", &TimedLifeComponentTYPEProxy))
-        return false;
-
-    if(engine->RegisterObjectType(
-           "AgentProperties", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(!bindComponentTypeId(
-           engine, "AgentProperties", &AgentPropertiesTYPEProxy))
-        return false;
-
-    if(engine->RegisterObjectMethod("AgentProperties",
-           "void setSpeciesName(string newString)",
-           asMETHOD(AgentProperties, setSpeciesName), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("AgentProperties",
-           "void setParentEntity(ObjectID parentId)",
-           asMETHOD(AgentProperties, setParentEntity), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("AgentProperties",
-           "void setAgentType(string newString)",
-           asMETHOD(AgentProperties, setAgentType), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("AgentProperties",
-           "string getSpeciesName()", asMETHOD(AgentProperties, getSpeciesName),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("AgentProperties",
-           "ObjectID getParentEntity()",
-           asMETHOD(AgentProperties, getParentEntity), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("AgentProperties", "string getAgentType()",
-           asMETHOD(AgentProperties, getAgentType), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-    return true;
-}
-
-template<class WorldType>
-bool
-    bindCellStageMethods(asIScriptEngine* engine, const char* classname)
-{
-
-    if(!Leviathan::BindStandardWorldMethods<CellStageWorld>(engine, classname))
-        return false;
-
-#include "generated/cell_stage_bindings.h"
-
-    ANGLESCRIPT_BASE_CLASS_CASTS_NO_REF(Leviathan::StandardWorld,
-        "StandardWorld", CellStageWorld, "CellStageWorld");
-
-    return true;
-}
-
-template<class WorldType>
-bool
-    bindMicrobeEditorMethods(asIScriptEngine* engine, const char* classname)
-{
-
-    if(!Leviathan::BindStandardWorldMethods<MicrobeEditorWorld>(
-           engine, classname))
-        return false;
-
-#include "generated/microbe_editor_bindings.h"
-
-    ANGLESCRIPT_BASE_CLASS_CASTS_NO_REF(Leviathan::StandardWorld,
-        "StandardWorld", CellStageWorld, "MicrobeEditorWorld");
 
     return true;
 }
@@ -1591,298 +362,6 @@ bool
 }
 
 bool
-    bindScriptAccessibleSystems(asIScriptEngine* engine)
-{
-    // ------------------------------------ //
-    // SpawnSystem
-    if(engine->RegisterFuncdef(
-           "ObjectID SpawnFactoryFunc(CellStageWorld@ world, Float3 pos)") <
-        0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectType("SpawnSystem", 0, asOBJ_REF | asOBJ_NOCOUNT) <
-        0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("SpawnSystem",
-           "void removeSpawnType(SpawnerTypeId spawnId)",
-           asMETHOD(SpawnSystem, removeSpawnType), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("SpawnSystem",
-           "SpawnerTypeId addSpawnType(SpawnFactoryFunc@ factory, double "
-           "spawnDensity, "
-           "double spawnRadius)",
-           asFUNCTION(addSpawnTypeProxy), asCALL_CDECL_OBJFIRST) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    // Process System
-    if(engine->RegisterObjectType(
-           "ProcessSystem", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("ProcessSystem",
-           "void setProcessBiome(int biomeId)",
-           asMETHOD(ProcessSystem, setProcessBiome), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("ProcessSystem",
-           "double getDissolved(CompoundId compoundData)",
-           asMETHOD(ProcessSystem, getDissolved), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-    // ------------------------------------ //
-    // CompoundCloudSystem
-    if(engine->RegisterObjectType(
-           "CompoundCloudSystem", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("CompoundCloudSystem",
-           "bool addCloud(CompoundId compound, float density, const Float3 &in "
-           "worldPosition)",
-           asMETHOD(CompoundCloudSystem, addCloud), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("CompoundCloudSystem",
-           "int takeCompound(CompoundId compound, const Float3 &in "
-           "worldPosition, float rate)",
-           asMETHOD(CompoundCloudSystem, takeCompound), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("CompoundCloudSystem",
-           "int amountAvailable(CompoundId compound, const Float3 &in "
-           "worldPosition, float rate)",
-           asMETHOD(CompoundCloudSystem, takeCompound), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    // ------------------------------------ //
-    // PlayerMicrobeControlSystem
-
-    // static
-    if(engine->SetDefaultNamespace("PlayerMicrobeControlSystem") < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterGlobalFunction(
-           "Float3 getTargetPoint(GameWorld &in worldWithCamera)",
-           asFUNCTION(PlayerMicrobeControlSystem::getTargetPoint),
-           asCALL_CDECL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->SetDefaultNamespace("") < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-
-    return true;
-}
-
-bool
-    registerPatches(asIScriptEngine* engine)
-{
-    // ------------------------------------ //
-    // Patch
-    ANGELSCRIPT_REGISTER_REF_TYPE("Patch", Patch);
-
-    if(engine->RegisterObjectBehaviour("Patch", asBEHAVE_FACTORY,
-           "Patch@ f(const string &in name, int32 id, const Biome &in "
-           "biomeTemplate)",
-           asFUNCTION(Patch::factory), asCALL_CDECL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("Patch", "const string& getName() const",
-           asMETHOD(Patch, getName), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("Patch", "int32 getId() const",
-           asMETHOD(Patch, getId), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("Patch", "bool addNeighbour(int32 id)",
-           asMETHOD(Patch, addNeighbour), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("Patch",
-           "Float2 getScreenCoordinates() const",
-           asMETHOD(Patch, getScreenCoordinates), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("Patch",
-           "void setScreenCoordinates(Float2 coordinates)",
-           asMETHOD(Patch, setScreenCoordinates), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    // TODO: Would be much safer to have reference counting for biomes
-    if(engine->RegisterObjectMethod("Patch", "const Biome@ getBiome() const",
-           asMETHODPR(Patch, getBiome, () const, const Biome&),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("Patch", "Biome@ getBiome()",
-           asMETHODPR(Patch, getBiome, (), Biome&), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("Patch",
-           "const Biome@ getBiomeTemplate() const",
-           asMETHODPR(Patch, getBiomeTemplate, () const, const Biome&),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("Patch",
-           ("bool addSpecies(Species@ species, int32 population = " +
-               std::to_string(INITIAL_SPECIES_POPULATION) + ")")
-               .c_str(),
-           asMETHOD(Patch, addSpeciesWrapper), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("Patch", "uint64 getSpeciesCount() const",
-           asMETHOD(Patch, getSpeciesCount), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("Patch",
-           "Species@ getSpecies(uint64 index) const",
-           asMETHOD(Patch, getSpeciesWrapper), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("Patch",
-           "int32 getSpeciesPopulation(const Species@ species) const",
-           asMETHOD(Patch, getSpeciesPopulationWrapper), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("Patch",
-           "array<int32>@ getNeighbours() const",
-           asFUNCTION(patchGetNeighboursWrapper), asCALL_CDECL_OBJFIRST) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    // ------------------------------------ //
-    // PatchMap
-    ANGELSCRIPT_REGISTER_REF_TYPE("PatchMap", PatchMap);
-
-    if(engine->RegisterObjectBehaviour("PatchMap", asBEHAVE_FACTORY,
-           "PatchMap@ f()", asFUNCTION(PatchMap::factory), asCALL_CDECL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("PatchMap", "Patch@ getCurrentPatch()",
-           asMETHOD(PatchMap, getCurrentPatchWrapper), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("PatchMap",
-           "int32 getCurrentPatchId() const",
-           asMETHOD(PatchMap, getCurrentPatchId), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("PatchMap", "Patch@ getPatch(int32 id)",
-           asMETHOD(PatchMap, getPatchWrapper), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("PatchMap", "bool addPatch(Patch@ patch)",
-           asMETHOD(PatchMap, addPatchWrapper), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("PatchMap",
-           "Species@ findSpeciesByName(const string &in name)",
-           asMETHOD(PatchMap, findSpeciesByNameWrapper), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-
-    mapWrapperTypeInfo =
-        Leviathan::ScriptExecutor::Get()->GetASEngine()->GetTypeInfoByDecl(
-            "array<const Patch@>");
-
-    if(!mapWrapperTypeInfo) {
-        LOG_ERROR("could not get type info for map wrapper");
-        return false;
-    }
-
-    if(engine->RegisterObjectMethod("PatchMap",
-           "array<const Patch@>@ getPatches() const",
-           asFUNCTION(patchMapGetPatchesWrapper), asCALL_CDECL_OBJFIRST) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("PatchMap", "array<Patch@>@ getPatches()",
-           asFUNCTION(patchMapGetPatchesWrapper), asCALL_CDECL_OBJFIRST) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    // ------------------------------------ //
-    // PatchManager
-    if(engine->RegisterObjectType(
-           "PatchManager", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("PatchManager", "PatchMap@ getCurrentMap()",
-           asMETHOD(PatchManager, getCurrentMapWrapper), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    return true;
-}
-
-bool
-    registerTimedWorldOperations(asIScriptEngine* engine)
-{
-    // ------------------------------------ //
-    // PatchManager
-    if(engine->RegisterObjectType(
-           "TimedWorldOperations", 0, asOBJ_REF | asOBJ_NOCOUNT) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("TimedWorldOperations",
-           "void onTimePassed(double timePassed)",
-           asMETHOD(TimedWorldOperations, onTimePassed), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterFuncdef("void ElapsedTimeFunc(GameWorld@ world, double "
-                               "elapsed, double totalTimePassed)") < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("TimedWorldOperations",
-           "void registerEffect(const string &in name, ElapsedTimeFunc@ "
-           "callback)",
-           asFUNCTION(registerEffectProxy), asCALL_CDECL_OBJFIRST) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    return true;
-}
-
-bool
     registerAutoEvo(asIScriptEngine* engine)
 {
     // ------------------------------------ //
@@ -1998,31 +477,6 @@ bool
 }
 
 bool
-    registerTweakedProcess(asIScriptEngine* engine)
-{
-    ANGELSCRIPT_REGISTER_REF_TYPE("TweakedProcess", TweakedProcess);
-
-    if(engine->RegisterObjectMethod("TweakedProcess",
-           "float get_tweakRate() const",
-           asMETHOD(TweakedProcess, getTweakRate), asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectMethod("TweakedProcess",
-           "float get_capacity() const", asMETHOD(TweakedProcess, getCapacity),
-           asCALL_THISCALL) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    if(engine->RegisterObjectProperty("TweakedProcess",
-           "const BioProcess process", asOFFSET(TweakedProcess, process)) < 0) {
-        ANGELSCRIPT_REGISTERFAIL;
-    }
-
-    return true;
-}
-
-bool
     thrive::registerThriveScriptTypes(asIScriptEngine* engine)
 {
     if(!registerLockedMap(engine))
@@ -2056,16 +510,13 @@ bool
     if(!bindThriveComponentTypes(engine))
         return false;
 
-    if(!bindScriptAccessibleSystems(engine))
-        return false;
-
-    if(!registerPlayerData(engine))
-        return false;
-
     if(!registerSimulationDataAndJsons(engine))
         return false;
 
-    if(!registerHexFunctions(engine))
+    if(!registerTweakedProcess(engine))
+        return false;
+
+    if(!registerOrganelles(engine))
         return false;
 
     if(!registerSpecies(engine))
@@ -2074,16 +525,19 @@ bool
     if(!registerPatches(engine))
         return false;
 
+    if(!bindScriptAccessibleSystems(engine))
+        return false;
+
+    if(!registerPlayerData(engine))
+        return false;
+
+    if(!registerHexFunctions(engine))
+        return false;
+
     if(!registerTimedWorldOperations(engine))
         return false;
 
     if(!registerAutoEvo(engine))
-        return false;
-
-    if(!registerTweakedProcess(engine))
-        return false;
-
-    if(!registerOrganelles(engine))
         return false;
 
     if(engine->RegisterObjectType("ThriveGame", 0, asOBJ_REF | asOBJ_NOCOUNT) <
@@ -2174,13 +628,7 @@ bool
     //     ANGELSCRIPT_REGISTERFAIL;
     // }
 
-
-
-    if(!bindCellStageMethods<CellStageWorld>(engine, "CellStageWorld"))
-        return false;
-
-    if(!bindMicrobeEditorMethods<MicrobeEditorWorld>(
-           engine, "MicrobeEditorWorld"))
+    if(!bindWorlds(engine))
         return false;
 
     if(engine->RegisterObjectMethod("ThriveGame",
