+++ src/scripting/script_initializer.h
@@ -14,4 +14,25 @@ bool
 bool
     registerOrganelles(asIScriptEngine* engine);
 
+bool
+    bindScriptAccessibleSystems(asIScriptEngine* engine);
+
+bool
+    bindWorlds(asIScriptEngine* engine);
+
+bool
+    bindThriveComponentTypes(asIScriptEngine* engine);
+
+bool
+    registerPatches(asIScriptEngine* engine);
+
+bool
+    registerTimedWorldOperations(asIScriptEngine* engine);
+
+bool
+    registerTweakedProcess(asIScriptEngine* engine);
+
+bool
+    registerSimulationDataAndJsons(asIScriptEngine* engine);
+
 } // namespace thrive
